Updated buffer process for accuracy (< .001 shifts) + additional metrics
New Rmd outputs the old specialized Ames metrics (still with the enhanced accuracy) along with a new file (ames_spatialized_indexed) with additional metrics regarding the makeup of the land cover in each buffer beyond percentage of cover (this is done via the landscape metrics package). It also produces the buffers themselves, for reuse, as well as images and summary statistics.

Area selection process has been enhanced to avoid needing entire county NLCD downloads and instead is replaced with an adjustable buffer for ease of area selection. Note: changing the area selection size and re-downloading may not appear in the plotting generation process without a restart of R [rm(list=ls()) has proven insufficient for this issue].

A final note on the "enhanced accuracy". While the process avoid unnecessary transformations to reduce internal rounding or degradation issues, it does not effect the fact that the original dataset uses approximated home locations. I do not attempt to correct for this. The process may more accurate, but the true underlying values are still subject to those original approximations and projection choices